requirements.txt
beautifulsoup4
google-generativeai
requests
